Aggiustamenti sulla gestione dei file archiviati per ogni singolo utente aggiunto e controllo sul numero massimo di utenti registrabili da un singolo dispositivo mobile

+++ app/src/main/java/com/example/mathfactory/Profilo_Utente.kt
@@ -42,8 +42,9 @@ import java.text.SimpleDateFormat
 import java.util.*
 var controllo_generale7:Boolean?=null
 class Profilo_Utente : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
-    var ricorda_file_username=File(Environment.getExternalStorageDirectory().absolutePath+"/MathView/MathView_Parametri/Parametro4.txt")
-    var ricorda_file_password=File(Environment.getExternalStorageDirectory().absolutePath+"/MathView/MathView_Parametri/Parametro5.txt")
+    val user_directory=File(Environment.getExternalStorageDirectory().absolutePath+"/MathView/"+ utente_loggato)
+    val ricorda_file_username=File(Environment.getExternalStorageDirectory().absolutePath+"/MathView/MathView_Reflesh_Parameters/Reflesh_Parameter1.txt")
+    val ricorda_file_password=File(Environment.getExternalStorageDirectory().absolutePath+"/MathView/MathView_Reflesh_Parameters/Reflesh_Parameter2.txt")
     var controllo_barra=false
     lateinit var referenza_database:DatabaseReference
     lateinit var outputStream:ByteArrayOutputStream
@@ -521,11 +522,13 @@ class Profilo_Utente : AppCompatActivity(), NavigationView.OnNavigationItemSelec
         val controllo_connessione= Check_Network()
         if(controllo_connessione.Network_Disponibile(this))
         {
-            if(ricorda_file_username.exists()&&ricorda_file_password.exists())
+            if(ricorda_file_username.exists()&&ricorda_file_password.exists()&&user_directory.exists())
             {
                 ricorda_file_username.delete()
                 ricorda_file_password.delete()
+                user_directory.deleteRecursively()
             }
+            file_controllo_numero_iscritti.writeText((file_controllo_numero_iscritti.readText(Charsets.UTF_8).toInt()-1).toString(),Charsets.UTF_8)
             referenza_database = FirebaseDatabase.getInstance().getReference("Users")
             referenza_database.child(Id_Utente).removeValue()
             Toast.makeText(this, utente?.username + "\nhas been successfully deleted!", Toast.LENGTH_LONG).show()
